fix(#1396): keep shadcn Select stock; wrap Radix portal outside ui/
There is no shadcn Select patch for the ItemText→SelectValue portal
(current radix-nova registry still uses empty SelectValue + bare
ItemText). Revert ui/select.tsx and put the workaround in an app-owned
wrapper so a future shadcn add does not clobber it.
