+++ src/components/billing/add-credits-dialog.tsx
@@ -25,7 +25,7 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '@/components/ui/select';
+} from '@/components/select';
 import { Skeleton } from '@/components/ui/skeleton';
 import {
   createCheckoutSessionFn,
